# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A proxy service for inspecting and forwarding API requests.",
        "title": "Inspect Proxy API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/chat/completions": {
            "post": {
                "description": "Forwards chat completion requests to the configured endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openai"
                ],
                "summary": "Forward chat completions requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/completions": {
            "post": {
                "description": "Forwards completion requests to the configured endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openai"
                ],
                "summary": "Forward completions requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/embeddings": {
            "post": {
                "description": "Forwards embedding requests to the configured endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openai"
                ],
                "summary": "Forward embeddings requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/moderations": {
            "post": {
                "description": "Forwards moderation requests to the configured endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openai"
                ],
                "summary": "Forward moderations requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ApiResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                }
            }
        }
    }
}